Use new build system: CMake (#34)
I added all new dependencies (for debian stretch or newer),
in the README. Swiften is now used from the system installation,
so I removed the part for compiling it. I also removed the
Troubleshooting part, because this is now automatically recognized
by cmake.

I removed the QMake project files, they're not needed anymore.

Android builds are a bit tricky with CMake, but I think we can
do that with the KDE ECM (extra cmake modules), as Kirigami.
Android is currently not supported.

This closes #18.