i18n: Add Russian translation file (#59)
You'll be able to translate it on weblate:
https://hosted.weblate.org/projects/kaidan/translations/
It is also a test if it really works. This commit also adds an example ts file,
if you want to add a new language, just copy it and do `make`, to update it.
I renamed the German ts file to only `de.ts` instead of `de_DE.ts`, because at
the moment we only have one translation for all German countries.