Add SQLite database for messages and roster caching
The new SQLite3 database is saved (on Linux) under:
~/.local/share/KaidanIM/kaidan/messages.sqlite3 (This is the AppData
location from Qt)

It currently has two tables one for the roster and one for messages.
The roster table has two coloumns: jid and name. The message table
has six: author, author_resource, recipient, recipient_resource,
timestamp and message. 'recipient' and 'author' contain a bare JID
(a JID without the resource) and the resource saved in
author/recipent_resource, if available. The resource is not really
necessary, but I thought it might be cool to add e.g. some stats for
this, later.

In the Kaidan.cpp/h, I moved the handleMessageReceived part into the
MessageController, later we should also move the handle Presence
Received into a PresenceController or something as this.

I also removed the 'import harbour.kaidan 1.0' in QML, because, now
we don't have any new QML-types from Kaidan -> you can't import it.

The Chat-GUI is still very ugly, we/I should definitely imporve it
in a later commit.

Closes #55.