CMake: Improve i18n build system, Make Kaidan installable (#66)
All translations are now bundled into the kaidan binary (@licensing@
only possible because translations and code are GPL), so we don't have
any problems locating the qm-files anymore. :)

The generation of the qm-files is more manual now, but so there are no
unncessary outputs on the command line anymore.

And best news: You can install kaidan now via. `make install`.