Remove ORGANIZATION_NAME, DOMAIN
They were added without reasons and then Qt thought Kaidan is installed
in /usr/share/KaidanIM/kaidan/ <- not what we want, so I removed it.
But this also makes all configs uncompatible. They're now in:
~/.config/kaidan/kaidan.conf and ~/.local/share/kaidan/messages.sqlite3

I hope that was the last time for the next time to change config names.