Roster: Sort using last exchanged
The new `lastExchanged` coloumn in the RosterModel contains
a ISODate string with the time when the last message has
arrived or has been sent. The roster is now sorted using this
property.

The message controller is now able to communicate with the roster
controller directly and updates the lastExchanged whenever a
message arrives/is been sent.

To preserve the lastExchanged property, the roster is now only
'updated' instead of being replaced every login. It looks up
which JIDs have been removed and added, adds/removes them
and updates the name of all the others (lastExchanged keeps the
same). Roster Versioning would be the better solution here and
also generally ...