ChatPage: New design showing time/isDelivered
This changes the design of the ChatPage or the underlying ChatMessage
object to show the date and a green checkmark if delivered.

The green checkmark is from Conversations (GPLv3).

The time format is 'dd MMM, hh:mm' / '31 Dec, 23:59', but it would
probably be better if this format would be changed dynamically, so that
all messages of the current day only show hours and minutes and messages
older than one year also show the year.

Closes #103, closes #104.