AvatarStorage: Deploy IDs dynamically to QML (#122)
This will make it possible to show automatically updating avatars of a
specific JID. Also the avatar hashes don't have to be saved in the
message database anymore, and so problems when changing accounts are
fixed now.

The AvatarStorage now has a signal, that is emitted when the avatar of a JID
was changed. Connected to it is the NOTIFY signal of the Q_PROPERTY of
the avatar storage, so all avatars are updated then.