ChatPage: Show real avatars using new RoundImage
This adds a new RoundImage QML type, that has a QtGraphicalEffect for
removing the corners and making it rounded (this is currently the only
way doing it, since an Image has no radius property as a Rectangle).

The avatar is now also a bit larger, I think it looks better so.
I removed mipmap filtering for the avatars on the ChatPage, because it
looked not good and a bit blurred on the avatars I tested it with. Maybe
I'll change that later again.