Bump copyright to 2018
This updates all copyright headers in the source files to 2018. Also,
there are not all contributors listed with a seperate copyright line,
only one line for all Kaidan developers and contributors with a 'link'
to a full list of contributors in the LICENSE file.

It would be very hard to maintain the old strategy and adding each
contributor to each file. Also the copyright lines were get too long then.

All Kaidan headings are uniform, now. I replaced 'A cross platform XMPP
client' with 'A user-friendly XMPP client for every device!'.