LICENSE: Add generation script and make machine-readable
This adds a script in utils/generate-license.py that will generate a new
LICENSE file upon the git history and information set in it. It's in the
debian copyright format, but it would be easily possible to use a
different style/format.

Since this will gather information from the git history, all
contributors will get their copyright line, even translators. This
should solve the problem with "see git history for more details" not
being a correct copyright for the participants in the git history.

In the script, there are replacements, and excludes of user ids. If
you've got several user ids in the git history, you can add replacements
for them, so they are not displayed several times in the LICENSE file.
Excludes are used for Weblate and anonymous translations.