Fix RichText labels in Plasma (Mobile)
This adds explicit textFormat settings in most labels, so labels are always
parsed, if wanted and never if not. So now, the messages aren't parsed anymore,
so the "send image hack" won't work anymore. This is also for security reasons,
because you probably don't want your system to always check any incoming links
in img tags.