Fix not connecting on the first time
This fixes that the server wasn't changed after the first connection. The
problem was that gloox::Client::setServer() only sets the server for connecting
to, not the JID that is used to for authenticating on the server. There's a
comment in the documentation that says to change the server manually by getting
a reference of the jid in the client, but that doesn't work since the returned
jid is marked as const.

The solution is to create a new class (GlooxClient) that will have another
function that is able to set the server in the JID.