Add Ubuntu Touch packaging & CI builds (#150)
This adds Ubuntu Touch click packaging and automated build scripts for Travis-CI builds. Building
is and will be only possible with CMake.

Currently the icons aren't working correctly (as in all CMake builds with a KF5 build from 5.40 to
5.43). There also might be still some problems with scaling, but the rest should work fine.

Publishing to the Open Store is planned, as soon as we get some more feedback.

Ubuntu Touch builds currently use a workaround for the application name: it'll use the application
id/uri for it, needed because of app armor. But I'd like to change to this principle on all platforms, so
in the future the configs and other persistant files will be saved in `~/.config/im.kaidan.kaidan` or
`~/.local/share/im.kaidan.kaidan/`, although I'm a bit unsure for normal desktop linux, there.

However, happy testing! :)