Fix release build TLS connection bug!
The problem was that the ServiceDiscoveryManager also is a
gloox::ConnectionHandler, but isn't return anything upon a handleTls call, so
this resulted in a true return in gcc (<= 7) in debug mode. gcc (<= 7) in
release mode and gcc-8 and clang in debug mode made it just fail. g++-8
and clang in release mode just gave a segfault.

Fixes #147.