travis: Fix that apt cache is empty at begin
Due to a change at travis-ci, apt update won't be executed at container start
up. Possibilities are to enable it in the travis.yml explicitly again or to do
it manually. Doing it manually is better for us, because we don't need another
apt update at start up for ubuntu touch builds.