avatars: Optimize vCard fetchment, Use SHA1 hashes locally
I made a mistake and somehow forgot to also check if the hash in the vCard based
avatars presence actually differs from the local hash saved for the JID. Thus,
before the vCards were fetched unnecessarily often. To do this I had to use
SHA1 hashes (instead of SHA3-256) for identifying the avatars.

Now the vCards only get fetched on start up, if there's no local avatar saved
for the JID, so if somebody has no avatar in their vCard the vCard will be
fetched every time at start up. Unfortunately this can't be further optimized,
but on the other hand that are just some bytes and I think that's totally ok.