database: Add message types and mediaUrl; Clean up
This introduces different message types (text, image, video, audio, document,
file). They should be used to determine the correct message element in the UI.

Also, there is a new mediaUrl in the Messages table which will be used to
save the HTTP urls to download the media shares.

The database version is increased to 5.