RosterPage: Add tooltips for presence information and JID
This adds tool tips to the roster. They show the alias/name of the contact,
their full JID (in small), the presence status type (i.e. 'Available', 'Away',
etc.) and the optional presence status message. The font of the status type is
in the apropriate color, so for 'Away' the font is orange.

Closes #77.