Rewrite RosterModel to fix all the problems
This makes the RosterModel simpler, fixes the display errors and should
improve the performance. If the database is directly modified (as before),
you also need to reload the complete database cache, which means executing
`select()`. This will take much time. The better way is to work with the
cache and using record(), setRecord() and removeRecord(), which will
directly update the ui. submitAll() is used to write all changes in the
cache to the database.

I actually don't know if I'm doing it right, but I'm sure it's much
better now than before. I only started to use all those SQL things
directly just because I had a look how the shmoose project does it.
Seems so they (geobra) have done it the wrong way ...

Now the same for the MessageModel ...