media-sharing: Display images in chat, Fix chat colour issues
Images are now displayed in the chat. For the case that they are
available locally (currently only possible by sending them), the images
are loaded directly, otherwise they're loaded via. HTTP. Of course that
still needs to be changed since those images are not cached and thus are
downloaded every time.

Fixes the `mediaLocation` attribute in the database of messages; it's
now the full file path, before it was accidently set to the folder's
path.

The colour for timestamp labels of incoming messages wasn't clearly
visible in some configurations, it is now always a bit darker (1.3
times) and well visible in all common styles.