EmojiModel: Fix crash in destructor
We can't really know in which order the QmlEngine destroyes the objects, so sometimes the Kaidan::instance() is already deleted when the the EmojiModel destructor is called.
The leads to favourite emojis being lost (oh no) and a segfault.

Fix it by saving when a favourite was added, not when the model is destroyed.