Make pending messages correctable
The message ID in the database describes the ID of the latest message
correction now (instead of the ID of original message). There's a new
column in the Messages table now for storing the 'replaceId', the ID of
the original message.

Little GUI change: Unavailable options in the message context menu are
invisible instead of being disabled.