PresenceCache: Rewrite most of the logic and simplify API for QML
This rewrites the logic behind the resource selection (or adds logic to
it at all).

The architecture behind the communication with QML is now also
different. Instead of manually creating a hacky binding from QML using
Q_INVOKABLEs and a changed signal, there's a new class doing that
automatically in C++. I'm not 100%-sure whether it's actually better
memory-wise, but it's definitely much easier to use.

The new QProperty system would really help here: We could avoid the
whole UserPresenceModel class then and directly return a QProperty for
the user...

The error presence is ignored for now. That requies some additional
special logic.

Fixes #308.