Ignore QXmppPresence::Invisible as it's unused and obsolete
Making the presence in the roster `visible: false` isn't a good idea
anyways. Also invisible presences can't be sent by the server (and so
never end up in our roster). "invisible" is only used by the client to
indicate to the server that it should hide its presence (aka making it
"unavailable").

Apart from that QXmpp also doesn't even follow the XEP: It uses
"invisible" as availability type instead of as presence type.