Move last message preview updates into RosterModel/Message
Since the last message and last exchanged values are calculated
dynaimcally now, we can move all this code into the Message where it
actually belongs. The updates are all done in one place now (the
RosterModel). This makes the mess a bit smaller and the code is not
duplicated in multiple locations anymore (MessageHandler /
RosterManager / UploadManager).

This also fixes the bug when the last message is loaded from the
database and the media url is displayed instead of the actual preview
text like "File" or "File: Hello World".