Search emojis only by shortcut ':' and short name
Searching emojis was cumbersome and the toggling search field moved the
whole emoji picker. The search is now opened by simply entering ':' and
the emoji's short name. The search is removed from the normal emoji
picker view.

That way, this power user feature can be accessed quicker. For average
users it makes the normal view simpler.