Restructure QML to XMPP Thread communication
Before this everything from QML that wanted to communicate with the XMPP
thread and the XMPP managers needed to use signals in Kaidan.h. The new
concept makes it possible to access the XMPP Managers directly although
you still need to use signals in order to communicate with the different
threads.