Update singleapp code to 4aeac8f
We will probably need to stay with the single application for a while,
because the other options like KDBusService probably won't work on all
platforms (like Windows and macOS).

There have been many bug-fixes and deprecation fixes regarding to Qt, so
updating is a good idea. I don't want to do this too frequently to not
mess up Kaidan's git history with it, but updating after 2.5 years seems
to be fine. :D