Swap message bubble colors
This is needed in preparation for the upcoming group chat
implementation. The message senders will be displayed inside of the
message bubble in the user's consistent colour.

By this the readability is improved and conflicting colors (e.g. blue
font on blue background) in terms of a pleasant user interface are
avoided.